backend/pyproject.toml
[project]
name = "fantasy-climbing-league-backend"
version = "0.1.0"
description = "Backend for Fantasy Climbing League"
authors = [
    {name = "FelixAbrahamsson",email = "felixabrahamsson1@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.125.0,<0.126.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "supabase (>=2.27.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
